Remove global CollisionGroup for a per-scene variant
